"refactor(#22): 라우팅 수정"
