+++ src/router/constants.js
@@ -1,0 +1,4 @@
+export const ROUTE = {
+  main: '/',
+  game: '/game',
+};
